+++ changelog.md
@@ -7,7 +7,7 @@
 
 #### v8.17
 
-- Updated error responses on all webapps and groot to always return a json with a standard format
+- Updated error responses on all webapps and groot to always return a json with a standard error format
 - Fixed kamehouse-shell scripts permissions
 - Fixed bugs